test
Migrating OnDynamicNetworkPreAndPostSpawn into its own class to avoid interfering with the generic tests.
